fix vite-plugin
