+++ packages/hono-takibi/src/generator/zod/z/integer.ts
@@ -1,4 +1,4 @@
-import type { Schema } from "../../../openapi/index.js"
+import type { Schema } from '../../../openapi/index.js'
 
 /**
  * Generates a Zod schema for integer types based on OpenAPI schema.
