+++ packages/hono-takibi/src/vite-plugin/index.ts
@@ -1,9 +1,13 @@
+import fsp from 'node:fs/promises'
 import path from 'node:path'
+import SwaggerParser from '@apidevtools/swagger-parser'
 import type { PluginOption, ViteDevServer } from 'vite'
-import { vite } from './vite.js'
-import { viteTsp } from './vite-tsp.js'
+import { fmt } from '../format/index.js'
+import { zodOpenAPIHono } from '../generator/zod-openapi-hono/openapi/index.js'
+import type { OpenAPI } from '../openapi/index.js'
+import { typeSpecToOpenAPI } from '../typespec/index.js'
 
-export default function HonoTakibiVite({
+export default async function HonoTakibiVite({
   input,
   output,
   exportType = true,
@@ -13,16 +17,42 @@ export default function HonoTakibiVite({
   output: `${string}.ts`
   exportType?: boolean
   exportSchema?: boolean
-}): PluginOption {
-  const run = async (): Promise<void> => {
+}): Promise<PluginOption> {
+  const run = async () => {
     if (
       typeof input === 'string' &&
       ((p: `${string}.yaml` | `${string}.json` | `${string}.tsp`): p is `${string}.tsp` =>
         p.endsWith('.tsp'))(input)
     ) {
-      await viteTsp(input, output, exportType, exportSchema)
+      try {
+        const tsp = await typeSpecToOpenAPI(input)
+        const openAPI = (await SwaggerParser.parse(tsp as unknown as string)) as OpenAPI
+        const hono = zodOpenAPIHono(openAPI, exportSchema, exportType)
+        const code = await fmt(hono)
+        if (!code.ok) {
+          console.error(`${code.error}`)
+          return false
+        }
+        await fsp.mkdir(path.dirname(output), { recursive: true })
+        await fsp.writeFile(output, code.value, 'utf-8')
+        return true
+      } catch (e) {
+        return false
+      }
     } else {
-      await vite(input, output, exportType, exportSchema)
+      try {
+        const openAPI = (await SwaggerParser.parse(input)) as OpenAPI
+        const hono = zodOpenAPIHono(openAPI, exportSchema, exportType)
+        const code = await fmt(hono)
+        if (!code.ok) {
+          console.error(`${code.error}`)
+          return false
+        }
+        await fsp.mkdir(path.dirname(output), { recursive: true })
+        await fsp.writeFile(output, code.value, 'utf-8')
+      } catch (e) {
+        return false
+      }
     }
   }
